update history page
